% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_filepath.R
\name{find_filepath}
\alias{find_filepath}
\title{Find where file paths occur in text
Finds the start and end positions of all file paths occurring in a text.}
\usage{
find_filepath(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
List of matrices
}
\description{
Find where file paths occur in text
Finds the start and end positions of all file paths occurring in a text.
}

